#5 implement rich text formatting
