+++ packages/core/src/types/item.js
@@ -1,0 +1,6 @@
+// @flow
+export type ItemBody = {};
+
+export type ItemFormat = {
+  layout: 'none' | 'grid' | 'flex',
+};
