+++ packages/editor/src/richtext/RichTextDesign.js
@@ -13,16 +13,16 @@ import {
   UnderlineButton,
   UnorderedListButton,
 } from 'draft-js-buttons';
-import { Box, Button as MuiButton } from '@material-ui/core';
+import { Box, IconButton } from '@material-ui/core';
 import type { RichTextBody, RichTextFormat } from '@seine/core';
 import { UPDATE_BLOCK_EDITOR, UPDATE_BLOCK_FORMAT } from '@seine/core';
 import { defaultDraftFormat } from '@seine/content';
+import { useAutoCallback } from 'hooks.macro';
 import styled from 'styled-components/macro';
 
-import theme from './RichTextDesign.module.css';
-import VerticalAlignTopButton from './VerticalAlignTopButton';
-import VerticalAlignCenterButton from './VerticalAlignCenterButton';
-import VerticalAlignBottomButton from './VerticalAlignBottomButton';
+import SidebarHeading from '../ui/SidebarHeading';
+import SidebarSection from '../ui/SidebarSection';
+
 import { defaultDraftEditor } from './RichTextEditor';
 
 type Props = {
@@ -30,17 +30,14 @@ type Props = {
   format: RichTextFormat,
 };
 
-const Button = styled(MuiButton)``;
-
-const DraftButton = ({ as: Button, className, ...props }) => (
-  <Button
-    {...props}
-    theme={{
-      ...theme,
-      button: `${className} ${theme.button}`,
-    }}
-  />
-);
+const DraftButton = styled(IconButton).attrs(({ className }) => ({
+  theme: { button: className },
+}))`
+  width: 2em;
+  svg {
+    fill: currentColor;
+  }
+`;
 
 /**
  * @description Rich text design panel.
@@ -54,87 +51,86 @@ export default function RichTextDesign({
   } = defaultDraftEditor,
   dispatch,
 }: Props) {
-  const asProps = {
+  const draftFormProps = {
     as: DraftButton,
 
     editorState,
-    getEditorState: React.useCallback(() => editorState, [editorState]),
-    setEditorState: React.useCallback(
-      (state) =>
-        dispatch({
-          type: UPDATE_BLOCK_EDITOR,
-          editor: { state },
-        }),
-      [dispatch]
+    getEditorState: useAutoCallback(() => editorState),
+    setEditorState: useAutoCallback((state) =>
+      dispatch({
+        type: UPDATE_BLOCK_EDITOR,
+        editor: { state },
+      })
     ),
     alignment:
       (format && format.textAlignment) || defaultDraftFormat.textAlignment,
-    setAlignment: React.useCallback(
-      ({ alignment: textAlignment }) =>
-        dispatch({
-          type: UPDATE_BLOCK_FORMAT,
-          format: { textAlignment },
-        }),
-      [dispatch]
-    ),
-  };
-
-  const verticalAlignmentAsProps = {
-    ...asProps,
-
-    alignment:
-      (format && format.verticalAlignment) ||
-      defaultDraftFormat.verticalAlignment,
-    setAlignment: React.useCallback(
-      ({ alignment }) =>
-        dispatch({
-          type: UPDATE_BLOCK_FORMAT,
-          format: { verticalAlignment: alignment },
-        }),
-      [dispatch]
+    setAlignment: useAutoCallback(({ alignment: textAlignment }) =>
+      dispatch({
+        type: UPDATE_BLOCK_FORMAT,
+        format: { textAlignment },
+      })
     ),
   };
 
   return (
-    <Box display={'flex'} flexWrap={'wrap'}>
-      {editorState && (
-        <>
-          <Box display={'flex'}>
-            <Button {...asProps} forwardedAs={BoldButton} />
-            <Button {...asProps} forwardedAs={ItalicButton} />
-            <Button {...asProps} forwardedAs={UnderlineButton} />
-          </Box>
+    <>
+      <SidebarHeading>Text</SidebarHeading>
+      <SidebarSection>
+        {editorState && (
+          <>
+            <Box display={'flex'}>
+              <Box display={'flex'} mr={3}>
+                <DraftButton {...draftFormProps} component={BoldButton} />
+                <DraftButton {...draftFormProps} component={ItalicButton} />
+                <DraftButton {...draftFormProps} component={UnderlineButton} />
+              </Box>
 
-          <Box display={'flex'}>
-            <Button {...asProps} forwardedAs={AlignBlockLeftButton} />
-            <Button {...asProps} forwardedAs={AlignBlockCenterButton} />
-            <Button {...asProps} forwardedAs={AlignBlockRightButton} />
-          </Box>
+              <Box display={'flex'}>
+                <DraftButton
+                  {...draftFormProps}
+                  component={HeadlineOneButton}
+                />
+                <DraftButton
+                  {...draftFormProps}
+                  component={HeadlineTwoButton}
+                />
+                <DraftButton
+                  {...draftFormProps}
+                  component={HeadlineThreeButton}
+                />
+              </Box>
+            </Box>
 
-          <Box display={'flex'}>
-            <Button
-              {...verticalAlignmentAsProps}
-              forwardedAs={VerticalAlignTopButton}
-            />
-            <Button
-              {...verticalAlignmentAsProps}
-              forwardedAs={VerticalAlignCenterButton}
-            />
-            <Button
-              {...verticalAlignmentAsProps}
-              forwardedAs={VerticalAlignBottomButton}
-            />
-          </Box>
+            <Box display={'flex'}>
+              <Box display={'flex'} mr={3}>
+                <DraftButton
+                  {...draftFormProps}
+                  component={OrderedListButton}
+                />
+                <DraftButton
+                  {...draftFormProps}
+                  component={UnorderedListButton}
+                />
+              </Box>
 
-          <Box display={'flex'}>
-            <Button {...asProps} forwardedAs={HeadlineOneButton} />
-            <Button {...asProps} forwardedAs={HeadlineTwoButton} />
-            <Button {...asProps} forwardedAs={HeadlineThreeButton} />
-            <Button {...asProps} forwardedAs={OrderedListButton} />
-            <Button {...asProps} forwardedAs={UnorderedListButton} />
-          </Box>
-        </>
-      )}
-    </Box>
+              <Box display={'flex'}>
+                <DraftButton
+                  {...draftFormProps}
+                  component={AlignBlockLeftButton}
+                />
+                <DraftButton
+                  {...draftFormProps}
+                  component={AlignBlockCenterButton}
+                />
+                <DraftButton
+                  {...draftFormProps}
+                  component={AlignBlockRightButton}
+                />
+              </Box>
+            </Box>
+          </>
+        )}
+      </SidebarSection>
+    </>
   );
 }
