+++ packages/editor/src/table/index.js
@@ -1,4 +1,3 @@
 // @flow
 export { default as TableEditor } from './TableEditor';
-export { default as TableToolbar } from './TableToolbar';
 export * from './constants';
